Bug Fixes
The item shop was using a players "UserGroup" attribute to keep track of
hunter or prop. ULX admin mod uses the "UserGroup" to keep track of the
players admin status. Every round it would remove ULX admin rights and
set it to hunter or prop. The point shop has been rewritten to use the
team's number instead.

Accessing the player model menu in the point shop is now disabled for
props. Ultimately we should limit only hunters to custom models so there
is no confusion.  This would also fix issues with the hunter halos.